feat: update user repository to use ULID type for IDs
Refactor user repository methods to accept ULID or string types for user IDs and actor IDs. This change enhances the flexibility of the repository by allowing the use of ULIDs, which are more suitable for distributed systems.

Also, update versioning in _version.py and add pydantic-extra-types dependency in pyproject.toml.